+++ CHANGELOG.md
@@ -13,6 +13,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Removed
 
+## [0.7.1](https://github.com/balance-io/balance-manager/releases/tag/0.7.1)
+
+### Changed
+* bugfix for send modal native output update
+
+## [0.7.0](https://github.com/balance-io/balance-manager/releases/tag/0.7.0)
+
+### Added
+* Added Greek language translation
+* Added WalletConnect support
+
+### Changed
+ 
+### Removed
+
 ## [0.6.1](https://github.com/balance-io/balance-manager/releases/tag/0.6.1)
 
 ### Added
@@ -27,8 +42,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Changed
 * Changed Shapeshift withdrawal list to default to first item instead of BNT
 * Changed default Ledger addresses shown to 20
-  
-### Removed
 
 ## [0.6.0](https://github.com/balance-io/balance-manager/releases/tag/0.6.0)
 
@@ -58,7 +71,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ## [0.5.4](https://github.com/balance-io/balance-manager/releases/tag/0.5.3)
 
 ### Added
-
 * Bookmark Reminder (anti-phishing) Ribbon
 * Dropdown on account view for choosing language
 * Link to Ledger support page on Ledger connection page